Fix the problem of never being schedulable caused by the failure of intermediate states in some schedulers

+++ cmd/device-monitor/main.go
@@ -52,7 +52,7 @@ func main() {
 		klog.Fatalf("Initialization of kubeConfig failed: %v", err)
 	}
 	kubeConfig, err := client.NewKubeConfig(
-		client.WithQPSBurst(float32(opt.QPS), opt.Burst),
+		client.WithQPSBurst(opt.QPS, opt.Burst),
 		client.WithDefaultUserAgent())
 	if err != nil {
 		klog.Fatalf("Create kubeConfig failed: %v", err)
